LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-js-innerhtml-expected.txt

PASS fileName(window.testdiv.children[0].currentSrc) is "image-set-4x.png"
PASS successfullyParsed is true

TEST COMPLETE
